Revert "Plan 9 from Bell Labs Fourth Edition 2005-12-12"
This reverts commit 22a127bbfe4dd304949cc596400de973c0138e31.

+++ sys/src/cmd/scat/scat.c
@@ -1288,7 +1288,7 @@ lookup(char *s, int doreset)
 		if(t == s)
 			goto BadCoords;
 		/* degree sign etc. is optional */
-		if((uchar)*t == L'°')
+		if(*t == L'°')
 			deg = DEG(getra(s));
 		if(doreset)
 			reset();
